refactor: move types to core
